requirements.txt
audiomentations==0.37.0
datasets==3.0.1
einops==0.8.0
librosa==0.10.2.post1
matplotlib==3.7.1
numpy==2.1.2
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.5.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
tqdm==4.66.5
transformers==4.44.2
